Refactor Optimizer class (#1180)
* Refactor Optimizer class

The main change is to serialize the optimizer state and not the
instance itself. Saving the instance restricted changes applied to the
class and the use of optimizer wrappers. This change is backward
compatible with existing checkpoints.

The learning rate management is also updated: the learning rate value
is always computed dynamically based on the initial learning
rate and the current training step. The learning rate decay schedules
are expected to be passed as callable taking a step and returning a
scaling factor.

* Fix step loaded from old checkpoints

* Use functools.partial instead of lambdas

* Use @classmethod instead of @staticmethod

* Some renames

* Move LR scheduling out of Adafactor optimizer